+++ xfer/background_publisher_test.go
@@ -1,0 +1,7 @@
+package xfer_test
+
+import "testing"
+
+func TestBackgroundPublisher(t *testing.T) {
+	t.Skip("TODO")
+}
