+++ xfer/http_publisher_test.go
@@ -1,7 +1,6 @@
 package xfer_test
 
 import (
-	"bytes"
 	"compress/gzip"
 	"encoding/gob"
 	"encoding/json"
@@ -13,7 +12,6 @@ import (
 	"time"
 
 	"github.com/gorilla/handlers"
-
 	"github.com/weaveworks/scope/report"
 	"github.com/weaveworks/scope/test"
 	"github.com/weaveworks/scope/xfer"
@@ -82,32 +80,3 @@ func TestHTTPPublisher(t *testing.T) {
 		t.Error("timeout")
 	}
 }
-
-func TestMultiPublisher(t *testing.T) {
-	var (
-		p              = &mockPublisher{}
-		factory        = func(string) (xfer.Publisher, error) { return p, nil }
-		multiPublisher = xfer.NewMultiPublisher(factory)
-	)
-
-	multiPublisher.Add("first")
-	if err := multiPublisher.Publish(&bytes.Buffer{}); err != nil {
-		t.Error(err)
-	}
-	if want, have := 1, p.count; want != have {
-		t.Errorf("want %d, have %d", want, have)
-	}
-
-	multiPublisher.Add("second") // but factory returns same mockPublisher
-	if err := multiPublisher.Publish(&bytes.Buffer{}); err != nil {
-		t.Error(err)
-	}
-	if want, have := 3, p.count; want != have {
-		t.Errorf("want %d, have %d", want, have)
-	}
-}
-
-type mockPublisher struct{ count int }
-
-func (p *mockPublisher) Publish(*bytes.Buffer) error { p.count++; return nil }
-func (p *mockPublisher) Stop()                       {}
